refactor: rename a file and move for parity between transports
