+++ packages/gateway/src/bot/decorators/Events.ts
@@ -1,8 +1,8 @@
 import { EventMetadataKey } from '@seedcord/core/internal';
 
-import type { ValidNonInteractionKeys } from '@handlers/BaseHandler';
 import type { EventHandler } from '@handlers/event';
 import type { EventFrequency } from '@miscellaneous/types';
+import type { ValidNonInteractionKeys } from '@src/handlers/interactionTypes';
 import type { ClientEvents } from 'discord.js';
 import type { Constructor } from 'type-fest';
