+++ packages/gateway/src/bot/defaults/UnhandledRepliable.ts
@@ -1,6 +1,6 @@
 import { InteractionHandler } from '@handlers/interaction/InteractionHandler';
 
-import type { Repliables } from '@handlers/BaseHandler';
+import type { Repliables } from '@src/handlers/interactionTypes';
 
 /**
  * @internal Default handler for unhandled repliable interactions.
