+++ packages/gateway/src/bot/gates/catalog/Nsfw.ts
@@ -4,8 +4,8 @@ import { pickNotice } from '@seedcord/core/internal';
 import { NotNsfw } from '@bot/notices';
 
 import type { InteractionGateContext } from '../Gate';
-import type { NonModalInteraction } from '@handlers/BaseHandler';
 import type { Gate, GateNoticeOptions } from '@seedcord/core';
+import type { NonModalInteraction } from '@src/handlers/interactionTypes';
 
 // a thread carries no nsfw flag of its own, so it inherits the parent channel's
 function channelIsNsfw(channel: InteractionGateContext['interaction']['channel']): boolean {
