+++ packages/gateway/src/bot/gates/runGates.ts
@@ -3,8 +3,8 @@ import { GuildMember } from 'discord.js';
 import { deriveEventActor } from '@miscellaneous/deriveEventActor';
 
 import type { EventGateContext, InteractionGateContext } from './Gate';
-import type { Repliables, ValidNonInteractionKeys } from '@handlers/BaseHandler';
 import type { Core } from '@interfaces/Core';
+import type { Repliables, ValidNonInteractionKeys } from '@src/handlers/interactionTypes';
 import type { ClientEvents } from 'discord.js';
 
 export function interactionGateContext(interaction: Repliables, core: Core): InteractionGateContext {
