+++ packages/gateway/src/handlers/BaseHandler.ts
@@ -1,32 +1,7 @@
 import { BaseHandler as CoreBaseHandler } from '@seedcord/core';
 
-import type { Core } from '@interfaces/Core';
-import type {
-    AnySelectMenuInteraction,
-    AutocompleteInteraction,
-    ButtonInteraction,
-    ChatInputCommandInteraction,
-    ClientEvents,
-    ContextMenuCommandInteraction,
-    Events,
-    ModalSubmitInteraction
-} from 'discord.js';
-
-export type ValidInteractionTypes =
-    | ChatInputCommandInteraction
-    | ButtonInteraction
-    | ModalSubmitInteraction
-    | AutocompleteInteraction
-    | AnySelectMenuInteraction
-    | ContextMenuCommandInteraction;
-
-export type ValidNonInteractionKeys = Exclude<keyof ClientEvents, Events.InteractionCreate>;
-
-export type ValidEventTypes = ValidInteractionTypes | ClientEvents[ValidNonInteractionKeys];
-
-export type Repliables = Exclude<ValidInteractionTypes, AutocompleteInteraction>;
-
-export type NonModalInteraction = Exclude<Repliables, ModalSubmitInteraction>;
+import type { Core } from '../interfaces';
+import type { ValidEventTypes } from './interactionTypes';
 
 /**
  * The gateway base, fixes the core base's `TCore` to the gateway {@link Core} and adds the `getEvent`
