+++ packages/gateway/src/handlers/index.ts
@@ -10,4 +10,4 @@ export {
 } from './interaction';
 export { EventHandler, EventMiddleware } from './event';
 export { RepliableHandler } from './RepliableHandler';
-export type { Repliables, ValidEventTypes, ValidInteractionTypes, ValidNonInteractionKeys } from './BaseHandler';
+export type { Repliables, ValidEventTypes, ValidInteractionTypes, ValidNonInteractionKeys } from './interactionTypes';
