+++ packages/gateway/src/handlers/interactionTypes.ts
@@ -1,0 +1,26 @@
+import type {
+    AnySelectMenuInteraction,
+    AutocompleteInteraction,
+    ButtonInteraction,
+    ChatInputCommandInteraction,
+    ClientEvents,
+    ContextMenuCommandInteraction,
+    Events,
+    ModalSubmitInteraction
+} from 'discord.js';
+
+export type ValidInteractionTypes =
+    | ChatInputCommandInteraction
+    | ButtonInteraction
+    | ModalSubmitInteraction
+    | AutocompleteInteraction
+    | AnySelectMenuInteraction
+    | ContextMenuCommandInteraction;
+
+export type ValidNonInteractionKeys = Exclude<keyof ClientEvents, Events.InteractionCreate>;
+
+export type ValidEventTypes = ValidInteractionTypes | ClientEvents[ValidNonInteractionKeys];
+
+export type Repliables = Exclude<ValidInteractionTypes, AutocompleteInteraction>;
+
+export type NonModalInteraction = Exclude<Repliables, ModalSubmitInteraction>;
