+++ packages/gateway/tests/bot/gates/runHandlerGates.test.ts
@@ -10,8 +10,8 @@ import { SlashHandler } from '@handlers/interaction/SlashHandler';
 
 import { TestNotice } from '../../utils/TestNotice';
 
-import type { Repliables } from '@handlers/BaseHandler';
 import type { Core } from '@interfaces/Core';
+import type { Repliables } from '@src/handlers/interactionTypes';
 import type { Events } from 'discord.js';
 
 declare module '@seedcord/core' {
