+++ packages/gateway/tests/bot/getConfirmation.test.ts
@@ -5,8 +5,8 @@ import { getConfirmation } from '@bot/confirm';
 import { CONFIRM_DEF } from '@bot/confirm/reserved';
 
 import type { DefaultConfirmOptions } from '@bot/confirm';
-import type { NonModalInteraction } from '@handlers/BaseHandler';
 import type { ReplyResponse } from '@seedcord/types';
+import type { NonModalInteraction } from '@src/handlers/interactionTypes';
 import type { ButtonInteraction, ModalSubmitInteraction } from 'discord.js';
 
 type Prompt = string | ((ids: { confirm: string; cancel: string }) => ReplyResponse);
